fix: Don't report too low time to merge if commits where re-written
Calculating time to merge as simply merged-first_commit produced invalid time to merge
statistics if commits in a PR where re-written during PR - e.g. if a workflow of fixups
and interactive rebase is used to apply review comments the first commit might actually
be much newer than the PR, leading to time-to-merge statistics shorter than the PR lead
time.

If a full time-to-merge can not be calculated (is lower than PR lead time) then lead time
is simply used, instead of reporting a wrong lower duration.